Edit Disabled For Spammers

+++ backend/schedule_auth_changer.php
@@ -38,12 +38,40 @@ $schedule = mysqli_real_escape_string($conn,$_POST['id']);
 $update_mode=mysqli_real_escape_string($conn,$_POST['mode']);
 $old_schedule=mysqli_real_escape_string($conn,$_POST['ref']);
 $error=false;
+
+$sql = "select * from schedule_request WHERE id=$schedule";
+$result = $conn->query($sql);
+if($result==FALSE)
+{
+    echo "ZERO";
+    die();
+}
+$row=$result->fetch_assoc();
+$updater=$row["user_id"];
+$bus=$row['bus_id'];
+
+
 if($state==0) { // Rejected
 
     $sql = "DELETE from `schedule_request` WHERE id=$schedule;";
     echo $sql;
     if ($conn->query($sql) == TRUE) {
         echo "ONE";
+        $sql="SELECT * from users WHERE id=$updater;";
+        $result = $conn->query($sql);
+        $row=$result->fetch_assoc();
+        $repu=$row["neg_repu"];
+        $repu=$repu+2;
+        $sql="UPDATE users SET neg_repu=$repu WHERE id=$updater;";
+        if ($conn->query($sql) == TRUE) {
+            //User Reputation Updated
+            echo "DONE";
+        }
+        else {
+            $error=true;
+            echo "ERR";
+            die();
+        }
     }
     else {
         $error=true;
@@ -54,16 +82,7 @@ if($state==0) { // Rejected
 //Bus	Trip Type	Time	Endpoint	Driver	Bus Number	Comment	User	User Reputation	Suggested at
 
 else { // Accepted
-    $sql = "select * from schedule_request WHERE id=$schedule";
-    $result = $conn->query($sql);
-    if($result==FALSE)
-    {
-        echo "ZERO";
-        die();
-    }
-    $row=$result->fetch_assoc();
-    $updater=$row["user_id"];
-    $bus=$row['bus_id'];
+
 
     /* Update Schedule Database */
     if($update_mode==0)//update
